added health check in install.sh

+++ install.sh
@@ -1,5 +1,35 @@
 #!/usr/bin/env bash
 
+if nvim --version | head -n1 | grep -vq 0.7; then
+  echo "expected nvim 0.7, aborting..."
+  exit 1
+fi
+
+if pip3 list | grep pynvim | grep -vq pynvim; then
+  echo "expected pynvim, aborting..."
+  exit 1
+fi
+
+if tmux -V | grep -vq 3.2; then
+  echo "expected tmux 3.2, aborting..."
+  exit 1
+fi
+
+if ! which rg > /dev/null; then
+  echo "rg not found, aborting..."
+  exit 1
+fi
+
+if ! which fdfind > /dev/null; then
+  echo "fdfind not found, aborting..."
+  exit 1
+fi
+
+if ! which fzf > /dev/null; then
+  echo "fzf not found, aborting..."
+  exit 1
+fi
+
 rm -rf \
   ~/.config/nvim/init.lua \
   ~/.config/nvim/lua \
